update query delete place to archive place

+++ server/src/models/place/place.queries.js
@@ -2,7 +2,8 @@ import mongoose from 'mongoose'
 
 import Place from './place.model'
 import '../inspecteur/inspecteur.model'
-import { appLogger } from '../../util'
+import { appLogger, techLogger } from '../../util'
+import { createArchivedPlaceFromPlace } from '../archived-place/archived-place.queries'
 
 export const PLACE_ALREADY_IN_DB_ERROR = 'PLACE_ALREADY_IN_DB_ERROR'
 
@@ -17,7 +18,20 @@ export const createPlace = async leanPlace => {
   return place.save()
 }
 
-export const deletePlace = async place => {
+export const deletePlace = async (place, reason, byUser, isCandilib) => {
+  if (!place) {
+    throw new Error('No place given')
+  }
+  try {
+    await createArchivedPlaceFromPlace(place, reason, byUser, isCandilib)
+  } catch (error) {
+    techLogger.error({
+      func: 'query-place-delete',
+      action: 'archive-place',
+      description: `Could not archive place { inspecteurId:${place.inspecteur}, centreId: ${place.centre}, date: ${place.date} }: ${error.message}`,
+      error,
+    })
+  }
   const deletedPlace = place
   await place.delete()
   return deletedPlace
